Sorting WetLoop to other Class and cleanings
